♻️Refactor : DetailView 생성하여 DetailViewController에서 분리
