+++ OpenMarket/OpenMarket/ViewController/DetailViewController.swift
@@ -12,14 +12,14 @@ protocol ProductUpdateDelegate: NSObject {
 }
 
 class DetailViewController: UIViewController {
+    lazy var detailView = DetailView(frame: view.frame)
     var product: Product?
     let numberFormatter: NumberFormatter = NumberFormatter()
     weak var delegate: ListUpdateDelegate?
     override func viewDidLoad() {
         super.viewDidLoad()
-        setupSubviewStructures()
-        setupLayoutConstraints()
-        configureContents()
+        detailView.configureContents(product: product)
+        self.view = detailView
         defineCollectionViewDelegate()
         guard let productName = product?.name else {
             return
@@ -31,168 +31,7 @@ class DetailViewController: UIViewController {
         let sheetButton = UIBarButtonItem(image: UIImage(systemName: "square.and.arrow.up"), style: .plain, target: self, action: #selector(requestSheet))
         self.navigationItem.rightBarButtonItem = sheetButton
 
-
-        collectionView.register(DetailCollectionViewCell.classForCoder(), forCellWithReuseIdentifier: "detailCell")
-    }
-    
-    let mainScrollView: UIScrollView = {
-        let scrollView = UIScrollView()
-        scrollView.isScrollEnabled = true
-        scrollView.showsHorizontalScrollIndicator = false
-        scrollView.showsVerticalScrollIndicator = true
-        scrollView.translatesAutoresizingMaskIntoConstraints = false
-        scrollView.contentInset = UIEdgeInsets(top: 0, left: 0, bottom: 0, right: 0)
-        return scrollView
-    }()
-    
-    let mainStackView: UIStackView = {
-        let stackView = UIStackView()
-        stackView.axis = .vertical
-        stackView.translatesAutoresizingMaskIntoConstraints = false
-        stackView.distribution = .fill
-        stackView.spacing = 5
-        stackView.isLayoutMarginsRelativeArrangement = true
-        stackView.layoutMargins = UIEdgeInsets(top: 0, left: 0, bottom: 0, right: 0)
-        return stackView
-    }()
-    
-    let nameLabel: UILabel = {
-        let label = UILabel()
-        label.translatesAutoresizingMaskIntoConstraints = false
-        label.font = .preferredFont(for: .title3, weight: .semibold)
-        return label
-    }()
-    
-    let stockLabel: UILabel = {
-        let label = UILabel()
-        label.translatesAutoresizingMaskIntoConstraints = false
-        label.textAlignment = .right
-        label.font = .preferredFont(forTextStyle: .title3)
-        label.textColor = .systemGray
-        return label
-    }()
-    
-    let priceLabel: UILabel = {
-        let label = UILabel()
-        label.translatesAutoresizingMaskIntoConstraints = false
-        label.font = .preferredFont(for: .body, weight: .semibold)
-        label.textAlignment = .right
-        return label
-    }()
-    
-    let bargainPriceLabel: UILabel = {
-        let label = UILabel()
-        label.translatesAutoresizingMaskIntoConstraints = false
-        label.textAlignment = .right
-        label.font = .preferredFont(for: .body, weight: .semibold)
-        return label
-    }()
-    
-    let descriptionLabel: UILabel = {
-        let label = UILabel()
-        label.translatesAutoresizingMaskIntoConstraints = false
-        label.numberOfLines = 0
-        return label
-    }()
-    
-    let informationStackView: UIStackView = {
-        let stackView = UIStackView()
-        stackView.axis = .horizontal
-        stackView.distribution = .equalSpacing
-        stackView.translatesAutoresizingMaskIntoConstraints = false
-        return stackView
-    }()
-    
-    var collectionView: UICollectionView = {
-        let collectionViewLayout = UICollectionViewFlowLayout()
-        collectionViewLayout.scrollDirection = .horizontal
-        collectionViewLayout.minimumLineSpacing = 20
-        let collectionView = UICollectionView(frame: .zero, collectionViewLayout: collectionViewLayout)
-        collectionView.translatesAutoresizingMaskIntoConstraints = false
-        collectionView.showsHorizontalScrollIndicator = true
-        collectionView.showsVerticalScrollIndicator = false
-        collectionView.allowsMultipleSelection = false
-        collectionView.decelerationRate = .fast
-        collectionView.contentInsetAdjustmentBehavior = .never
-        
-        return collectionView
-    }()
-    
-    func setupSubviewStructures() {
-        
-        informationStackView.addArrangedSubview(nameLabel)
-        informationStackView.addArrangedSubview(stockLabel)
-        
-        mainStackView.addArrangedSubview(collectionView)
-        mainStackView.addArrangedSubview(informationStackView)
-        
-        mainStackView.addArrangedSubview(priceLabel)
-        mainStackView.addArrangedSubview(bargainPriceLabel)
-        mainStackView.addArrangedSubview(descriptionLabel)
-        mainScrollView.addSubview(mainStackView)
-        self.view.addSubview(mainScrollView)
-        
-    }
-    
-    func setupLayoutConstraints() {
-        mainScrollView.topAnchor.constraint(equalTo: self.view.topAnchor).isActive = true
-        mainScrollView.bottomAnchor.constraint(equalTo: self.view.bottomAnchor).isActive = true
-        mainScrollView.leadingAnchor.constraint(equalTo: self.view.leadingAnchor, constant: 15).isActive = true
-        mainScrollView.trailingAnchor.constraint(equalTo: self.view.trailingAnchor, constant: -15).isActive = true
-        mainScrollView.centerXAnchor.constraint(equalTo: self.view.centerXAnchor).isActive = true
-        mainScrollView.centerYAnchor.constraint(equalTo: self.view.centerYAnchor).isActive = true
-        
-        mainStackView.topAnchor.constraint(equalTo: mainScrollView.topAnchor).isActive = true
-        mainStackView.bottomAnchor.constraint(equalTo: mainScrollView.bottomAnchor).isActive = true
-        mainStackView.leadingAnchor.constraint(equalTo: mainScrollView.leadingAnchor).isActive = true
-        mainStackView.trailingAnchor.constraint(equalTo: mainScrollView.trailingAnchor).isActive = true
-        mainStackView.centerXAnchor.constraint(equalTo: mainScrollView.centerXAnchor).isActive = true
-        
-        collectionView.leadingAnchor.constraint(equalTo: mainStackView.leadingAnchor).isActive = true
-        collectionView.trailingAnchor.constraint(equalTo: mainStackView.trailingAnchor).isActive = true
-        collectionView.heightAnchor.constraint(equalTo: self.view.heightAnchor, multiplier: 0.45).isActive = true
-        
-        informationStackView.leadingAnchor.constraint(equalTo: mainStackView.leadingAnchor).isActive = true
-        informationStackView.trailingAnchor.constraint(equalTo: mainStackView.trailingAnchor).isActive = true
-        
-        nameLabel.leadingAnchor.constraint(equalTo: informationStackView.leadingAnchor).isActive = true
-        stockLabel.trailingAnchor.constraint(equalTo: informationStackView.trailingAnchor).isActive = true
-        
-        priceLabel.trailingAnchor.constraint(equalTo: mainStackView.trailingAnchor).isActive = true
-        bargainPriceLabel.trailingAnchor.constraint(equalTo: mainStackView.trailingAnchor).isActive = true
-    }
-    
-    func configureContents() {
-        guard let product = product else {
-            return
-        }
-        
-        guard let currency = product.currency?.rawValue else {
-            return
-        }
-        
-        nameLabel.text = product.name
-        stockLabel.text = "남은 수량 : \(numberFormatter.numberFormatString(for: Double(product.stock)))"
-        priceLabel.text = "\(currency) \(numberFormatter.numberFormatString(for:product.price))"
-        
-        if product.discountedPrice != .zero {
-            guard let price = priceLabel.text else {
-                return
-            }
-            priceLabel.textColor = .red
-            priceLabel.attributedText = setTextAttribute(of: price, attributes: [.strikethroughStyle: NSUnderlineStyle.single.rawValue])
-            bargainPriceLabel.text = "\(currency) \(numberFormatter.numberFormatString(for:product.bargainPrice))"
-        }
-        
-        
-        descriptionLabel.text = product.description
-    }
-    
-    private func setTextAttribute(of target: String, attributes: [NSAttributedString.Key: Any]) -> NSAttributedString {
-        let attributedText = NSMutableAttributedString(string: target)
-        attributedText.addAttributes(attributes, range: (target as NSString).range(of: target))
-        
-        return attributedText
+        detailView.collectionView.register(DetailCollectionViewCell.classForCoder(), forCellWithReuseIdentifier: "detailCell")
     }
     
     @objc private func requestCancel() {
@@ -251,7 +90,7 @@ class DetailViewController: UIViewController {
                 self.deleteProduct(id: product.id, secret: data)
             case .failure(_):
                 DispatchQueue.main.async {
-                    self.showAlert(alertTitle: "비밀번호 실패")
+                    self.showAlert(alertTitle: "비밀번호가 맞지 않습니다.")
                 }
             }
         })
@@ -298,8 +137,8 @@ extension DetailViewController: UICollectionViewDelegate, UICollectionViewDataSo
     }
     
     private func defineCollectionViewDelegate() {
-        self.collectionView.delegate = self
-        self.collectionView.dataSource = self
+        detailView.collectionView.delegate = self
+        detailView.collectionView.dataSource = self
     }
     
     func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
@@ -309,8 +148,8 @@ extension DetailViewController: UICollectionViewDelegate, UICollectionViewDataSo
 
 extension DetailViewController: UIScrollViewDelegate {
     func scrollViewWillEndDragging(_ scrollView: UIScrollView, withVelocity velocity: CGPoint, targetContentOffset: UnsafeMutablePointer<CGPoint>) {
-        guard let flowLayout = self.collectionView.collectionViewLayout as? UICollectionViewFlowLayout else { return }
-        let cellWidth = collectionView.frame.width
+        guard let flowLayout = self.detailView.collectionView.collectionViewLayout as? UICollectionViewFlowLayout else { return }
+        let cellWidth = detailView.collectionView.frame.width
         let spaceBetweenCell = flowLayout.minimumLineSpacing
         let cellWidthWithSpace = cellWidth + spaceBetweenCell
         
@@ -331,7 +170,7 @@ extension DetailViewController: ProductUpdateDelegate {
             case .success(let data):
                 self.product = data
                 DispatchQueue.main.async {
-                    self.configureContents()
+                    self.detailView.configureContents(product: data)
                 }
             case .failure(_):
                 DispatchQueue.main.async {
